pyproject.toml
[tool.poetry]
name = "gips"
version = "0.1.0"
description = "generador de informes psicológicos"
authors = ["Héctor Urbina <hurbinas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
requests = "^2.31.0"
django-json-widget = "^1.1.1"
django-reversion = "^5.0.12"
beautifulsoup4 = "^4.12.3"
weasyprint = "^61.2"
rut-chile = "^2.0.1"
qrcode = {extras = ["pil"], version = "^7.4.2"}
gunicorn = "^21.2.0"
dj-database-url = "^2.1.0"
django-countries = "^7.6.1"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
pytest-django = "^4.8.0"
psycopg2-binary = "^2.9.9"
babel = "^2.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
